still a lot of to do - no failed to find id

+++ src/main/java/vortex/imwp/repositories/TaxRateRepository.java
@@ -6,6 +6,6 @@ import vortex.imwp.models.Country;
 import vortex.imwp.models.TaxRate;
 
 @Repository
-public interface TaxRateRepository extends JpaRepository<TaxRate, Integer> {
+public interface TaxRateRepository extends JpaRepository<TaxRate, Long> {
     TaxRate findByCountry(Country.Name country);
 }
